chore(pr-recommendations): made the fluxeditor readOnly and moved some things around
